moves all tests into a testing directory and splits the tests into separate files
