+++ .coveragerc
@@ -1,5 +1,15 @@
 [run]
-omit = /usr/*, gunicorn_config.py, deploy.py, app_test.py, app.py, pysite/websockets.py, pysite/views/*__init__.py, pysite/route_manager.py
+omit = /usr/*,
+       gunicorn_config.py,
+       deploy.py,
+       app_test.py,
+       app.py,
+       pysite/websockets.py,
+       pysite/views/*__init__.py,
+       pysite/route_manager.py,
+       pysite/migrations/runner.py
 
 [report]
-exclude_lines = return jsonify
+exclude_lines = return jsonify,
+                raise RuntimeError,
+                return
